Currently, if the result of an XQuery query is an empty sequence and if the output mode is one of  "xml", "xml-mapi", "xml-soap", "xml-mapi-soap", the following will be printed:
    <?xml version="1.0" encoding="utf-8"?>
    <XQueryResult>

    </XQueryResult>

Note the empty line between <XQueryResult> and </XQueryResult>.  This
output does not represent empty result, instead, it represent result
with *one* text node, with a '\n' as its content.  (In case the query
result is not empty, this '\n' is needed separate </XQueryResult> from
the last item of the query result.)

To prevent this unnecessary '\n' been printed, an extra parameter "bit
emptySeq" has been added to the "seqEnd" functions.  This parameter is
only used by xml_seqEnd to decide if it should print a '\n' before
</XQueryResult> or not.

Also added test for this bug.