removed the -s hacks for Xquery!
New use of MapiClient options (for Xquery only)
	-t 	enables as before tracing Mapi
	-targ	When -t gets some argument it will instead pass the tracing
		to the Xquery server (ie the old debug/tracing output)
	-T	Both mapi timing and xquery profiling
	-oformat This new flag can be used to change the output format of
		xquery. For example dm (default) and xml.

These options are passed to the server with the special prefix 'X'.
Normal queries are now prefixed with 's' or 'S' (last is a multiline statement)

The -i option for interactive is normally not used for the xquery language,
this now changes. Using
	MapiClient -i docname -lxquery < some_file.xml
should shred the document and add it to the persistent collection.