optimize functions created using sql (fixes a problem in functions of the form 'return (select * from tablex)'
function/aggr now can get module and implementation names, ie we can store
that some functions in 5 come from strange modules.
This feature isn't used jet... But will soon I hope. That should fix
the problems with the math functions in m5/sql.

bitwise and,or,xor,not are now named bit_{and,or,xor,not}. This
because and,or,not are sql keywords.