fixing BUG #1568652 "BATpropcheck/BATjoin":
in case BATfetchjoin was called with a result size estimation != oid_nil,
a result BAT of the given size was allocated and filled withou further checks;
without estimation (estimate == oid_nil) the size of the left input was used
as save upper bound.

Fix: in case an estimation < BATcount(l) is provided, we increase it to the
save upper bound of estimate = BATcount(l) (and hence still avoid size checking
in BATfetchjoin's inner loop.

Also added a respective test.