Implemented the `document management' type.
There is now a special primitive type in Pathfinder's type
system: `docmgmt'.  This is the return type of all document
management functions (add-doc(), del-doc()).  Type checking
will ensure that document management is not mixed with (a)
update statements (that update the content of some document,
but not the document collection itself) or (b) ordinary
XQuery results.