A change which was supposed to be smaller, but turned out to be quite large...  But testing runs fine on it, so all should be well.
Triggered by Niels' SQL code doing a strcmp on exception strings to
figure out in one case if an assertion was hit, it seemed like a good
idea to have a function to return the exception as an enum type, such
that the strcmp would be redundant, given that distinguising exception
names would be used.

So, this commit is exactly that change.  But, it turned out that the
jungle had to be trimmed down a bit, so the whole mal_exception.mx file
has been rewritten now (and looks nice and clean), with some functions
which have descriptive and suggestive names.  These have been propagated
through the total M5 code.

The exception structure itself hasn't changed, the way they are created
is.  A function now takes care of crafting the message, which takes an
code from an enum, controlling the exceptions available in the system.