removed declared and set but never used variable "unsigned int i" sql_column_name_new() global scope; thanks ti icc for finding and reporting this bug.
variable i is only used the the two for loops,
where it is locally declared as PFalg_att_t and PFalg_simple_type_t, respectively,
but appears to be used as unsigned int ... ?

Manuel,

is this "inconsistency" intended and/or is it save/correct to declare
i as PFalg_simple_type_t / PFalg_att_t but use it as unsigned int?

Please check, and modify the code if necessary.
Recall, for statements can only have one local declaration
(that, as here, can obviously declare and initialize multiple variables,
 but all of the same type).

Similar situation occur in multiple places, at least in this file.