Test for the new implementation.
All the features supported for functions are now supported for procedures. However, procedure and functions differ in two aspects (well, maybe one):
Procedure do not have return type and no return statements in their runtime_body.

A new field to the struct sql_func was added to simply the distinguish between t
hem and also to help in the dependencies check.