Implementation of the XRPC wrapper for external XQuery engines, as described in our paper submitted to VLDB07:
 The XRPC wrapper is a SOAP service handler that stores the incoming
 SOAP XRPC request messages in a temporary location, generates an XQuery
 query for this request, and executes it on an XQuery processor.

 The generated query is crafted to compute the result of a Bulk XRPC by
 calling the requested function on the parameters found in the message,
 and to generate the SOAP response message in XML using element
 construction.

This XRPC wrapper works with Saxon8-9J and Galax 0.7.2 .
To use the warpper:

java -jar <path-to>/xrpcwrapper-0.1.jar \
     --command "<command-to-run-your-xquery-engine>"

The "command" option is mandatory, since the wrapper has to know how to
run the XQuery engine.

For example, to use Saxon:
java -jar jars/xrpcwrapper-1.0.jar \
     --command "java -cp <path-to>/saxon8.jar net.sf.saxon.Query"

To use Galax:
java -jar jars/xrpcwrapper-1.0.jar \
     --command "<path-to>/galax-run"

The string value of the "command" option is passed literally to a Java
Proc to be executed.  If any options need to be passed to the XQuery
engine, they should be added to the string.

Examples can be found in the Tests sub-directory
(nl/cwi/monetdb/xrpc/Tests).  The module definition file and the XML
documents needs to be copied to "/tmp" manaully.