First step to have partition distribution. The distribution is done at the partition fragment level. Each fragment resulted from a bat partition can be located at a different location than the actual bat location.
For the moment the location information is (by default):
        host (localhost)
        port (50000)
        dbname (demo)
        user (monetdb)
        passwd (monetdb)

In the bpm_part structure a pointer to this structure was added. By default it points for the
default location.

Functions:
Function to set location:
	If the location does not exit it creates a new one.

Function to remove location:
	It only sets the location pointer equal NULL (maybe default would be better)
	However, the location is not removed from the location catalog because several
fragements might share the same location

Storage management:
	I only save the location id. When the fragment is loaded I use this id
to get its location information (if location catalog is not loaded yet it will be).
	The addition and remotion of the fragment location is also implemented


Note: It was compiled with strict compilation and no errors showed up. For the moment
I only started the mserver with option "--set sql_debug=16" to activate the partition
model and no errors showed up.