convention is to use: addHost(), setHost()
Add commands to reconnect to a remote host
Note: It is not a re-implementation of merser.reconnect. It uses
the location id to get the data for then call the mserver.reconnect
If the old location structure used by the reconnect can be re-created
this command could then be removed.

Add command to send fragment. However, it is not finished because we still need to
implement mserver.send();