Implemented the -w (wait) flag for start, stop and kill.
While the flag is not really necessary for the order of things
(merovingian will queue a client to cover the startup time), it may be
nice to just know when a database is really started, or stopped.  When
-w is used, also diagnostic messages can be printed when for instance
starting a database failed due to a crash.

This commit includes (or better said, superseedes) the following commits
from the stable branch, and hence those should NOT be propagated (they
will conflict for sure anyway):
sql/src/backends/monet5 monetdb.mx, SQL_2-22, 1.7.2.2, 1.7.2.3
sql/src/backends/monet5 monetdb.mx, SQL_2-22, 1.7.2.1, 1.7.2.2