Fix for bug #1886326: MonetDB JDBC driver exception for XQuery with no results
2 and a half hour later ...
Due to missing history (cvs moves) and missing documentation (on my
end) it took quite some while to figure out what was the "correct" way
to solve this bug.  In the end it al boiled down on the limitation of
the XQuery backend to only support full-results-at-once, which means the
resource allocation of JDBC needs to allocate the entire result at once.
If the result is 0 rows big, however, then things went wrong when
calculating the current cache block.  Anyway, for the future we now know
why we did this cache block calulation thing like this again...