The generic recycler framework. The recycler optimizer marks all variables that are potential useful for recycling. This applies to expressions over constants and side-effect free operations. The instruction is also marked as relevant to be monitored. The recycle version is set such that it become automatically initialized.
The interpreter has been wrapped with calls to the recycler, looking
for a cheap way to get the result. If not, the instruction is executed
and the result values become available for recycling.
Currently, all values are retained. The timing and other properties
are now available for more precise policies.

Note, the recycler works on identical execution of instructions only.
For example, optimization of select calls against stored results
should still be handled in a specialized operation, which redirects
the search against the base table to one already kept.