This checkin changes the handling of mmapped files with the different bat access modes and persistency. THis used to be handled via the heap filename and extra X.priv files. Its now handled with in the heap structure field storage and the bats newstorage fields together with a X.new file (allways write the new mmapped private file, and never release the old X file when it could be used).
The bulk of this changes was done by Peter. I just fixed some obvious
problems.

The default case (malloced bats) and the case where bats are
'read only' and forcefully writen is correctly handled.  These case
are mostly used. The remap-case was special (and still is), so
we don't expect problems with sql and xquery. These changes do fix
a reported data corruption problem.

It seem (as shown by the test results) to almost work.
However there are some problems.
	1) void heaps sometimes get materialized within a transaction.
	   ie we introduce a heap which wasn't there. THis together with
	   mmap must be check for correctness.
	2) unloading/loading gives problems, ie we don't (jet) correctly
	   check on loading if the storage needs to change from mmap->priv