How to spend an entire Sunday.  Either my sabatical turned me into a n00b, or the code is still the same maze it used to be.  (It is for a change nice compared to Python's hell code...)
remote appears to have been added, but not approved before, so this
commit includes the approval of that, for I like it, remote deserves to
be there by default.

I introduced a new atom, rmtobj, which is a small wrapper around a
string.  This allowed me to "see" when such remote identifier is used in
a remote.exec call.  Instead of pushing a string to the remote site,
with an identifier, the identifier name itself only has to be used in
the remote query.

This allows to perform e.g. a slice on a remote BAT, that you once put
there, which you can only reference via the rmtobj, which used to be a
string.  See also remote04.mal which tests this goofy feature.

Romulo might like this feature (a lot, actually).