Fixes to logging.
Pathfinder now no longer calls subcommit with BATs that belong to
the logger.  Instead, pathfinder starts and ends a logger transaction,
but with extra non-logger BATs added to the logger's subcommit.  This
means that there is an API change that affects MonetDB, MonetDB4 and
pathfinder.

pf_checkpoint was changed to start and end a logger transaction.  But
since sometimes a transaction needs to be started before pf_checkpoint
is called, it has an extra parameter which says whether to start a
transaction (and if so also get and release the logger lock).

A number of problems have been fixed in the logger as well.  When a
snapshot BAT is deleted from the logger catalog, it is removed from
the snapshots BAT and inserted into the freed BAT.  (This used to
happen only when a new version of the BAT was added again, it now also
happens when deleting.)  This freed BAT contains BATs that need to be
decref'ed later.  This now happens at logger restart instead of at
transaction end, since upon reading the log at server restart, the
BATs should still exist.

When the log is read at server restart, BATs that are logged as
snapshots should be handled the same as when they were initially added
or deleted.  This means that la_bat_use needs to add the BAT to the
snapshot BAT

This fixes bug [ 1976341 ] XQ: leftovers after deleting document (but
not bug [ 2009556 ] XQ: "Zombie" document in collection).