Fixed a number of leaks in the bat cache: - When a bat is freed that has capacity, say, 300, the bat was added   to the cache.  If a later invocation of the query needs a bat of   capacity 300, the cache was *not* consulted since it could only   guarantee bats with capacity 256.  If such a query needs and frees   many of those bats, and if that query is repeated many times, the   cache just grows. - When a query is repeated several times in different client sessions,   the chances that any temporary bats were reused is actually very   small (if they happen to ask for the correct size, see previous   bullet) since there were 64 different caches each serving a distinct   set of clients.  This means that the number of bats in the cache can   become 64 times the number that would actually be needed. - When doing fetchjoins on a str bat, tricks were being played that   cause a bat to be taken from the cache (if there was one of the   right size and in the right cache--see the previous bullets) and   then the type of the bat would be changed to str and the bat would   be deposited in the cache in a different bin.  This means that the   cache just accumulates str bats.
I fixed all these problems by redefining the cache buckets to be
buckets for bats of a particular capacity and by rounding up the
requested capacity to a multiple of BATTINY (256).  In this way, if a
query needs a bat of capacity 300, the capacity is rounded up to 512,
and a bat is found in the cache (if the cache isn't empty) and later
that bat is deposited in the same bucket, ready for reuse.

The number of size buckets is initially set to BATCACHE_BUCKETS (2,
i.e. capacities up to 512).  The size can be changed using MIL
bbp_batcache_minsize().  If set to 0, the cache is disabled.  There is
no MAL interface: that could be added in the development branch.
There is a built-in maximum for the number of buckets of
BATCACHE_MAXBUCKETS (4 i.e. capacities up to 2096).

The fetchjoin is changed to not play the "string trick" for small
bats, i.e. bats that could be allocated from the cache.

Since the cache is shared by all threads, there is more reuse of bats.
The flip side is contention among threads for access to the cache.
This is alleviated somewhat by having caches for different sizes.
However, there is definitely more potential for contention in the new
scheme, but there is also more potential for reuse (and more actual
reuse).

Also in this change: two new functions for debugging purposes:
BBPdump() and BBPdumpcache().  Neither is called anywhere in the code,
so they can only be called from a debugger session.  In the
development version we might add an interface to call these functions
from MAL/MIL.  Note however that neither of these functions is thread
safe.