Thanks to jhyde for the suggestion, Connection.setReadonly() indeed throws the infamous "writable mode not supported" message.
If seem to have interpreted once that this method should bail out if one
requests to set it to false (setReadOnly(false)), however, in the light
of apache-dbcp (which Mondrian uses), and the very careful wording of           the contract of setReadOnly, it seems we better see a call to this
function at most as a suggestion, hence we now fire off a warning when
one attempts to set read-only to false.

This makes Mondrian very (very very) happy.