fixed bug in rel_join_order. Again rewrite join's which have on one of the sides a referenced relation. (fixes problems with xquery/q01.sql).
This triggered problems, which were related to too eager 'dce' for referenced
relations.
The rel2bin_basetable now only returns the statements with names coming
from the expression list (if available). Before we returned both expressions
and the base column list, which lead to incorrect column usage.

Cleanup the print_stmt a bit